requirements.txt
annotated-types==0.6.0
anyio==4.3.0
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
distro==1.9.0
exceptiongroup==1.2.0
google-api-core==2.17.1
google-api-python-client==2.119.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-speech==2.25.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.62.0
grpcio-status==1.62.0
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httpx==0.27.0
idna==3.6
install==1.3.5
openai==1.12.0
pocketsphinx==5.0.3
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.2
pydantic_core==2.16.3
pydub==0.25.1
pyparsing==3.1.1
pytube==15.0.0
requests==2.31.0
rsa==4.9
sniffio==1.3.0
sounddevice==0.4.6
SpeechRecognition==3.10.1
tqdm==4.66.2
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.1
